requirements.txt
aistudio-sdk==0.3.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
attrdict==2.0.1
babel==2.18.0
bce-python-sdk==0.9.60
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==7.0.0
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cssselect==1.4.0
cssutils==2.11.1
ctranslate2==4.7.1
cycler==0.12.1
Cython==3.2.4
decorator==5.1.1
et_xmlfile==2.0.0
fastapi==0.128.4
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-babel==4.0.0
fonttools==4.61.1
fsspec==2026.2.0
future==1.0.0
h11==0.14.0
hf-xet==1.2.0
httpcore==1.0.7
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.10
ImageIO==2.37.2
imagesize==1.4.1
imgaug==0.4.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
keyboard==0.13.5
kiwisolver==1.4.9
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
modelscope==1.34.0
more-itertools==10.8.0
networkx==3.4.2
nvidia-cublas-cu12==12.3.4.1
nvidia-cuda-nvrtc-cu12==12.6.85
nvidia-cuda-runtime-cu12==12.3.101
nvidia-cudnn-cu12==9.0.0.312
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nvjitlink-cu12==12.6.85
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless
numpy==1.26.4
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==26.0
paddleocr==2.7.3
pillow==11.1.0
premailer==3.10.0
prettytable==3.17.0
protobuf==5.29.3
psutil==7.2.2
py-cpuinfo==9.0.0
pyclipper==1.4.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.3.0
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.14.3
rarfile==4.2
regex==2026.1.15
requests==2.32.5
ruamel.yaml==0.19.1
sacremoses==0.1.1
safetensors==0.7.0
scikit-image==0.26.0
scipy==1.17.0
sentencepiece==0.2.1
shapely==2.1.2
shellingham==1.5.4
shiboken6==6.10.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
termcolor==3.3.0
tifffile==2026.1.28
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.1.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.40.0
visualdl==2.5.3
wcwidth==0.6.0
Werkzeug==3.1.5